Improve Code Organization in React Components
