+++ src/components/analytics/analyticsHeader.tsx
@@ -1,5 +1,16 @@
 import { useNavigate } from "react-router";
 
+function HeaderStat({ label, value }: { label: string; value: number | string }) {
+    return (
+        <div className="flex flex-col gap-2.5">
+            <div className="font-mono text-xs text-teal-400 uppercase">{label}</div>
+            <div className="flex items-center justify-center font-mono text-lg font-bold">
+                {value}
+            </div>
+        </div>
+    );
+}
+
 export default function AnalyticsHeader({
     title,
     totalDataUsage,
@@ -23,22 +34,8 @@ export default function AnalyticsHeader({
             </button>
             <div className="w-1/2 font-mono text-lg text-teal-400">{title}</div>
             <div className="flex flex-1 items-center justify-evenly">
-                <div className="flex flex-col gap-2.5">
-                    <div className="font-mono text-xs text-teal-400 uppercase">
-                        {"Total Data Used"}
-                    </div>
-                    <div className="flex items-center justify-center font-mono text-lg font-bold">
-                        {totalDataUsage}
-                    </div>
-                </div>
-                <div className="flex flex-col gap-2.5">
-                    <div className="font-mono text-xs text-teal-400 uppercase">
-                        {"Videos Watched"}
-                    </div>
-                    <div className="flex items-center justify-center font-mono text-lg font-bold">
-                        {numVideosWatched}
-                    </div>
-                </div>
+                <HeaderStat label="Total Data Used" value={totalDataUsage} />
+                <HeaderStat label="Videos Watched" value={numVideosWatched} />
             </div>
         </div>
     );
