+++ src/components/analytics/weekUsage.tsx
@@ -9,6 +9,7 @@ import {
 import { useEffect, useState } from "react";
 import AnalyticsHeader from "./analyticsHeader";
 import AnalyticsBody from "./analyticsBody";
+import PageLayout from "./pageLayout";
 
 function getWeekTotalUsage(weekUsage: UsageByDay) {
     let total = 0;
@@ -53,14 +54,14 @@ export default function WeekUsage() {
 
     return (
         <>
-            <div className="flex h-screen w-full flex-col">
+            <PageLayout>
                 <AnalyticsHeader
                     title={formatDate(getLast7Days())}
                     totalDataUsage={formatBytes(getWeekTotalUsage(weekUsage))}
                     numVideosWatched={getNumVideosWatched(weekUsage)}
                 />
                 <AnalyticsBody usage={weekUsage} />
-            </div>
+            </PageLayout>
         </>
     );
 }
