+++ src/components/infoCard.tsx
